refactor(controller): split engines god-router into concern registrars
routes.ts carried recipe CRUD, launch lifecycle, model downloads, and
the runtime-management surface in one 387-line file. It becomes a
12-line aggregator over recipe-routes / lifecycle-routes /
download-routes / runtime-routes, with the shared observed-process
closure extracted to a factory. Handler bodies verbatim; registration
order preserved.
